+++ run_debug.py
@@ -3,7 +3,7 @@
 from cah.settings.config import Development
 
 if __name__ == '__main__':
-    Development.build_db_engine()
+    Development().build_db_engine()
     from cah.app import create_app
 
     app = create_app(config_class=Development, props=Development.SECRETS)